# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSK Smart Learning API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  name: string;
  currentLevel: number;
  targetLevel: number;
  onboardingCompleted: boolean;
  totalXp: number;
  dailyGoalMinutes: number;
  createdAt: string;
}

export interface UserProfileInput {
  name?: string;
  targetLevel?: number;
  dailyGoalMinutes?: number;
}

export interface OnboardingInput {
  name: string;
  targetLevel: number;
  placementScore?: number;
  recommendedLevel: number;
  dailyGoalMinutes?: number;
}

export type LessonType = typeof LessonType[keyof typeof LessonType];


export const LessonType = {
  pinyin_basics: 'pinyin_basics',
  dialogue: 'dialogue',
  vocabulary: 'vocabulary',
  grammar: 'grammar',
  review: 'review',
} as const;

export interface Lesson {
  id: number;
  level: number;
  unit: number;
  lessonNumber: number;
  title: string;
  titleChinese: string;
  type: LessonType;
  isLocked: boolean;
  isCompleted: boolean;
  xpReward: number;
  estimatedMinutes: number;
}

export interface DialogueLine {
  speaker: string;
  chinese: string;
  pinyin: string;
  translation: string;
}

export interface VocabularyItem {
  id: number;
  chinese: string;
  pinyin: string;
  translation: string;
  tone: number;
  /** @nullable */
  strokeOrder?: string | null;
  exampleSentence: string;
  exampleTranslation: string;
}

export type GrammarPointExamplesItem = {
  chinese: string;
  pinyin: string;
  translation: string;
};

export interface GrammarPoint {
  pattern: string;
  explanation: string;
  examples: GrammarPointExamplesItem[];
}

export interface LessonDetail {
  id: number;
  level: number;
  unit: number;
  lessonNumber: number;
  title: string;
  titleChinese: string;
  type: string;
  dialogue: DialogueLine[];
  vocabulary: VocabularyItem[];
  grammarPoints: GrammarPoint[];
  shadowingText: string;
  shadowingPinyin: string;
}

export interface DailyTask {
  lessonId: number;
  title: string;
  type: string;
  estimatedMinutes: number;
  isCompleted: boolean;
  order: number;
}

export interface DailyPlan {
  date: string;
  totalMinutes: number;
  completedMinutes: number;
  tasks: DailyTask[];
  motivationMessage: string;
}

export type LessonProgressStep = typeof LessonProgressStep[keyof typeof LessonProgressStep];


export const LessonProgressStep = {
  dialogue: 'dialogue',
  vocabulary: 'vocabulary',
  shadowing: 'shadowing',
  grammar: 'grammar',
  quiz: 'quiz',
} as const;

export interface LessonProgress {
  lessonId: number;
  step: LessonProgressStep;
  isCompleted: boolean;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  completedAt?: string | null;
}

export type LessonProgressInputStep = typeof LessonProgressInputStep[keyof typeof LessonProgressInputStep];


export const LessonProgressInputStep = {
  dialogue: 'dialogue',
  vocabulary: 'vocabulary',
  shadowing: 'shadowing',
  grammar: 'grammar',
  quiz: 'quiz',
} as const;

export interface LessonProgressInput {
  step: LessonProgressInputStep;
  isCompleted: boolean;
  /** @nullable */
  score?: number | null;
}

export type ProgressSummaryWeeklyXpItem = {
  day: string;
  xp: number;
};

export interface ProgressSummary {
  currentLevel: number;
  currentUnit: number;
  completedLessons: number;
  totalLessons: number;
  levelProgressPercent: number;
  vocabularyLearned: number;
  totalXp: number;
  weeklyXp: ProgressSummaryWeeklyXpItem[];
}

export interface StreakData {
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  lastStudyDate?: string | null;
  studiedToday: boolean;
}

export interface ShadowingInput {
  lessonId: number;
  audioBase64?: string;
  targetText: string;
  targetPinyin: string;
}

export type SyllableFeedbackColor = typeof SyllableFeedbackColor[keyof typeof SyllableFeedbackColor];


export const SyllableFeedbackColor = {
  green: 'green',
  yellow: 'yellow',
  red: 'red',
} as const;

export interface SyllableFeedback {
  syllable: string;
  pinyin: string;
  score: number;
  color: SyllableFeedbackColor;
  /** @nullable */
  errorType?: string | null;
}

export interface ShadowingResult {
  overallScore: number;
  syllableFeedback: SyllableFeedback[];
  tips: string[];
  passed: boolean;
}

export type QuizQuestionType = typeof QuizQuestionType[keyof typeof QuizQuestionType];


export const QuizQuestionType = {
  matching: 'matching',
  pinyin_input: 'pinyin_input',
  listening_choice: 'listening_choice',
  tone_selection: 'tone_selection',
  translation_choice: 'translation_choice',
} as const;

export interface QuizQuestion {
  id: number;
  type: QuizQuestionType;
  question: string;
  /** @nullable */
  options?: string[] | null;
  correctAnswer: string;
  /** @nullable */
  audioText?: string | null;
}

export type QuizType = typeof QuizType[keyof typeof QuizType];


export const QuizType = {
  mini_quiz: 'mini_quiz',
  checkpoint: 'checkpoint',
} as const;

export interface Quiz {
  id: number;
  /**
     * null for checkpoint exams
     * @nullable
     */
  lessonId?: number | null;
  level: number;
  type: QuizType;
  questions: QuizQuestion[];
  totalQuestions: number;
  /** @nullable */
  timeLimit?: number | null;
}

export interface QuizGenerateInput {
  lessonId: number;
  questionCount?: number;
}

export type QuizSubmitInputAnswersItem = {
  questionId: number;
  answer: string;
};

export interface QuizSubmitInput {
  quizId: number;
  answers: QuizSubmitInputAnswersItem[];
}

export type QuizResultQuestionResultsItem = {
  questionId: number;
  isCorrect: boolean;
  correctAnswer: string;
  userAnswer: string;
};

export interface QuizResult {
  quizId: number;
  score: number;
  correctCount: number;
  totalCount: number;
  passed: boolean;
  xpEarned: number;
  questionResults: QuizResultQuestionResultsItem[];
}

export type DashboardStatsWeeklyActivityItem = {
  day: string;
  xp: number;
  minutes: number;
};

export interface DashboardStats {
  totalXp: number;
  currentLevel: number;
  targetLevel: number;
  levelProgressPercent: number;
  vocabularyLearned: number;
  completedLessons: number;
  currentStreak: number;
  todayMinutes: number;
  dailyGoalMinutes: number;
  weeklyActivity: DashboardStatsWeeklyActivityItem[];
}

export type GetLessonsParams = {
level?: number;
unit?: number;
};