# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HSK Smart Learning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lessons:
    get:
      operationId: getLessons
      tags: [lessons]
      summary: Get lessons for a HSK level
      parameters:
        - in: query
          name: level
          schema:
            type: integer
          required: false
        - in: query
          name: unit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson detail with dialogue, vocabulary, grammar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
...
  /progress/lessons/{lessonId}:
    post:
      operationId: updateLessonProgress
      tags: [progress]
      summary: Record progress for a lesson step
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonProgressInput"
      responses:
        "200":
          description: Updated lesson progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"
...
  /quiz/generate:
    post:
      operationId: generateQuiz
      tags: [quiz]
      summary: Generate a mini-quiz for a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizGenerateInput"
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quiz/submit:
    post:
      operationId: submitQuiz
      tags: [quiz]
      summary: Submit quiz answers and get results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmitInput"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /quiz/checkpoint/{level}:
    get:
      operationId: getCheckpoint
      tags: [quiz]
      summary: Get HSK-format checkpoint exam for a level
      parameters:
        - in: path
          name: level
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checkpoint exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
...
components:
  schemas:
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        level:
          type: integer
        unit:
          type: integer
        lessonNumber:
          type: integer
        title:
          type: string
        titleChinese:
          type: string
        type:
          type: string
          enum: [pinyin_basics, dialogue, vocabulary, grammar, review]
        isLocked:
          type: boolean
        isCompleted:
          type: boolean
        xpReward:
          type: integer
        estimatedMinutes:
          type: integer
      required:
        - id
        - level
        - unit
        - lessonNumber
        - title
        - titleChinese
        - type
        - isLocked
        - isCompleted
        - xpReward
        - estimatedMinutes
    DialogueLine:
      type: object
      properties:
        speaker:
          type: string
        chinese:
          type: string
        pinyin:
          type: string
        translation:
          type: string
      required:
        - speaker
        - chinese
        - pinyin
        - translation
    VocabularyItem:
      type: object
      properties:
        id:
          type: integer
        chinese:
          type: string
        pinyin:
          type: string
        translation:
          type: string
        tone:
          type: integer
        strokeOrder:
          type: ["string", "null"]
        exampleSentence:
          type: string
        exampleTranslation:
          type: string
      required:
        - id
        - chinese
        - pinyin
        - translation
        - tone
        - exampleSentence
        - exampleTranslation
    GrammarPoint:
      type: object
      properties:
        pattern:
          type: string
        explanation:
          type: string
        examples:
          type: array
          items:
            type: object
            properties:
              chinese:
                type: string
              pinyin:
                type: string
              translation:
                type: string
            required:
              - chinese
              - pinyin
              - translation
      required:
        - pattern
        - explanation
        - examples
    LessonDetail:
      type: object
      properties:
        id:
          type: integer
        level:
          type: integer
        unit:
          type: integer
        lessonNumber:
          type: integer
        title:
          type: string
        titleChinese:
          type: string
        type:
          type: string
        dialogue:
          type: array
          items:
            $ref: "#/components/schemas/DialogueLine"
        vocabulary:
          type: array
          items:
            $ref: "#/components/schemas/VocabularyItem"
        grammarPoints:
          type: array
          items:
            $ref: "#/components/schemas/GrammarPoint"
        shadowingText:
          type: string
        shadowingPinyin:
          type: string
      required:
        - id
        - level
        - unit
        - lessonNumber
        - title
        - titleChinese
        - type
        - dialogue
        - vocabulary
        - grammarPoints
        - shadowingText
        - shadowingPinyin
...
    LessonProgress:
      type: object
      properties:
        lessonId:
          type: integer
        step:
          type: string
          enum: [dialogue, vocabulary, shadowing, grammar, quiz]
        isCompleted:
          type: boolean
        score:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
      required:
        - lessonId
        - step
        - isCompleted
    LessonProgressInput:
      type: object
      properties:
        step:
          type: string
          enum: [dialogue, vocabulary, shadowing, grammar, quiz]
        isCompleted:
          type: boolean
        score:
          type: ["integer", "null"]
      required:
        - step
        - isCompleted
...
    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [matching, pinyin_input, listening_choice, tone_selection, translation_choice]
        question:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: string
        audioText:
          type: ["string", "null"]
      required:
        - id
        - type
        - question
        - correctAnswer
    Quiz:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: ["integer", "null"]
          description: null for checkpoint exams
        level:
          type: integer
        type:
          type: string
          enum: [mini_quiz, checkpoint]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        totalQuestions:
          type: integer
        timeLimit:
          type: ["integer", "null"]
      required:
        - id
        - level
        - type
        - questions
        - totalQuestions
    QuizGenerateInput:
      type: object
      properties:
        lessonId:
          type: integer
        questionCount:
          type: integer
      required:
        - lessonId
    QuizSubmitInput:
      type: object
      properties:
        quizId:
          type: integer
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              answer:
                type: string
            required:
              - questionId
              - answer
      required:
        - quizId
        - answers
    QuizResult:
      type: object
      properties:
        quizId:
          type: integer
        score:
          type: integer
        correctCount:
          type: integer
        totalCount:
          type: integer
        passed:
          type: boolean
        xpEarned:
          type: integer
        questionResults:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              isCorrect:
                type: boolean
              correctAnswer:
                type: string
              userAnswer:
                type: string
            required:
              - questionId
              - isCorrect
              - correctAnswer
              - userAnswer
      required:
        - quizId
        - score
        - correctCount
        - totalCount
        - passed
        - xpEarned
        - questionResults
...