# lib/api-zod/src/generated/types/dialogueLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSK Smart Learning API
 * OpenAPI spec version: 0.1.0
 */

export interface DialogueLine {
  speaker: string;
  chinese: string;
  pinyin: string;
  translation: string;
}