# lib/api-zod/src/generated/types/grammarPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSK Smart Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { GrammarPointExamplesItem } from './grammarPointExamplesItem';

export interface GrammarPoint {
  pattern: string;
  explanation: string;
  examples: GrammarPointExamplesItem[];
}