# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSK Smart Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from './quizQuestion';
import type { QuizType } from './quizType';

export interface Quiz {
  id: number;
  /**
     * null for checkpoint exams
     * @nullable
     */
  lessonId?: number | null;
  level: number;
  type: QuizType;
  questions: QuizQuestion[];
  totalQuestions: number;
  /** @nullable */
  timeLimit?: number | null;
}