# lib/api-zod/src/generated/types/vocabularyItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSK Smart Learning API
 * OpenAPI spec version: 0.1.0
 */

export interface VocabularyItem {
  id: number;
  chinese: string;
  pinyin: string;
  translation: string;
  tone: number;
  /** @nullable */
  strokeOrder?: string | null;
  exampleSentence: string;
  exampleTranslation: string;
}